requirements.txt
attrdict==2.0.0
certifi==2018.4.16
chardet==3.0.4
cytoolz==0.9.0.1
Django==2.0.4
eth-abi==1.0.0
eth-account==0.1.0a2
eth-hash==0.1.2
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.0
eth-utils==1.0.2
flake8==3.5.0
gunicorn==19.7.1
hexbytes==0.1.0
idna==2.6
lru-dict==1.1.6
mccabe==0.6.1
psycopg2-binary==2.7.4
py-solc==2.1.0
pycodestyle==2.3.1
pycryptodome==3.6.1
pyflakes==1.6.0
pytz==2018.4
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
web3==4.1.0
websockets==4.0.1
whitenoise==3.3.1
